+++ src/test/kotlin/br/com/gamemods/regionmanipulator/RegionTest.kt
@@ -1,9 +1,8 @@
 package br.com.gamemods.regionmanipulator
 
+import org.junit.Assert
 import org.junit.Test
 import java.io.File
-import java.util.zip.ZipFile
-import java.util.zip.ZipInputStream
 
 class RegionTest {
     @Test
@@ -32,6 +31,22 @@ class RegionTest {
 
         val mca = RegionIO.readRegion(tempFile, RegionPos(-1, -2))
         RegionIO.writeRegion(tempFile, mca)
+    }
+
+    @Test(expected = CorruptChunkException::class)
+    fun testIssue2fix() {
+        val tempFile = File.createTempFile("r.-1,-1,", ".mca")
+        tempFile.deleteOnExit()
+        RegionTest::class.java.getResourceAsStream("/issue.2.r.-1.-1.mca").use { input ->
+            tempFile.outputStream().use { output ->
+                input.copyTo(output)
+            }
+        }
+        val region = RegionIO.readRegion(tempFile, RegionPos(-1, -1))
+        val chunkPos = ChunkPos(-24, -17)
+        Assert.assertNotNull(region.getCorrupt(chunkPos))
+
+        region[chunkPos]
     }
 
     @Test